Go to absurd length to enable arrow keys & Enter
Seriously what was I even thinking. It brought me back to Full Sail days
fiddling with jQuery this much but eventually I figured it out I think.

The autocomplete suggestions are now navigable with arrow keys (can't
wait to hear the Vim-ers complain :-p) and submittable with the Enter
key which is a lot better.